Integration_1/CloudStorage_Service/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
boto3==1.24.6
python-dotenv==0.19.2
python-multipart
requests